django_pg_backend/core/requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8

# Authentication
djangorestframework-simplejwt>=5.2.0

# Resume Parsing - PDF/DOCX Extraction
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0 
numpy>=2.2.6

# NLP - spaCy for Named Entity Recognition (Optional - Falls back to regex if not installed)
# spaCy requires Visual Studio C++ Build Tools on Windows
# Install from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# spacy>=3.7.0
# en-core-web-lg>=3.7.1

# Text Processing
regex>=2023.0.0

# Date Parsing
python-dateutil>=2.8.0

# Phone Number Parsing (Optional)
phonenumbers>=8.13.0

# Email Validation (Optional)
email-validator>=2.1.0

# Additional Utilities
pytz>=2023.3
